requirements.txt
Flask==3.1.0
PyJWT==2.10.1
requests==2.32.3
gunicorn==23.0.0
cryptography==44.0.3
